+++ packages/http-controllers/index.d.ts
@@ -1,5 +1,6 @@
 import "@egomobile/http-server";
-import type { IControllersOptions, IControllersResult } from "./src/types/index.js";
+import type { IControllersOptions, IControllersResult, ImportValues } from "./src/types/index.js";
+import type { Nilable } from "./src/types/internal.js";
 
 declare module "http" {
     interface IncomingMessage {
@@ -14,11 +15,12 @@ declare module "@egomobile/http-server" {
         /**
          * Initializes the underlying server instance, using controllers.
          *
-         * @param {string} [rootDir=controllers] The custom root directory of the controller files. Relative paths will be mapped to the current working directory.
+         * @param {string} [rootDir="controllers"] The custom root directory of the controller files. Relative paths will be mapped to the current working directory.
          * @param {IControllersOptions} options Custom options.
+         * @param {Nilable<ImportValues>} [imports] Additional imports.
          */
         controllers(): Promise<IControllersResult>;
-        controllers(rootDir: string): Promise<IControllersResult>;
+        controllers(rootDir: string, imports?: Nilable<ImportValues>): Promise<IControllersResult>;
         controllers(options: IControllersOptions): Promise<IControllersResult>;
     }
 }